tsuite/requirements.txt
click>=8.0.0
pyyaml>=6.0
jsonpath-ng>=1.5.0
docker>=6.0.0
flask>=2.0.0
requests>=2.28.0
rich>=13.0.0
